requirements.txt
diffusers==0.30.3
matplotlib==3.7.5
Pillow==10.4.0
torch==2.4.1+cu124
transformers==4.44.2
